requirements.txt
# 웹 프레임워크
fastapi>=0.95.0
uvicorn[standard]>=0.21.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.6

# OpenAI 및 LLM 관련
openai>=1.0.0
tiktoken>=0.4.0
langchain>=0.0.200
langchain-community>=0.0.10
langchain-openai>=0.0.2

# 벡터 데이터베이스
chromadb>=0.4.10
sentence-transformers>=2.2.2

# 문서 처리
unstructured>=0.7.0
pypdf>=3.15.0
docx2txt>=0.8
beautifulsoup4>=4.12.0
markdown>=3.4.0

# 음성 처리
pydub>=0.25.1
elevenlabs>=0.2.20

# 데이터베이스 
sqlalchemy>=2.0.0
pymysql>=1.1.0
alembic>=1.11.0

# 웹 클라이언트
httpx>=0.24.0
aiohttp>=3.8.5

# 인증
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# 유틸리티
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0
jinja2>=3.1.2

# 테스트
pytest>=7.3.0
pytest-asyncio>=0.21.0

# 개발 도구
black>=23.3.0
isort>=5.12.0
mypy>=1.2.0
ruff>=0.0.262 